Finalise doc updates
