+++ doc/Makefile
@@ -58,7 +58,7 @@ html:
 	# the embedding of images more robust
 	rm -rf $(BUILDDIR)/html/_images
 	#rm -rf _build/doctrees/
-	$(SPHINXBUILD) -b html $(ALLSPHINXOPTS) $(BUILDDIR)/html
+	$(SPHINXBUILD) -b html $(ALLSPHINXOPTS) $(BUILDDIR)/html -j auto
 	python3 fix_search.py  # TODO remove when sphinx_rtd_theme fixes upstream
 	touch $(BUILDDIR)/html/.nojekyll
 	@echo
